Split `sail-common` crate
