+++ crates/sail-common-datafusion/src/display.rs
