+++ crates/sail-common/Cargo.toml
@@ -3,21 +3,22 @@ name = "sail-common"
 version = { workspace = true }
 edition = { workspace = true }
 
+# DO NOT add DataFusion dependencies for this crate.
+# All utilities that depend on DataFusion should be defined in the `sail-common-datafusion` crate.
+
 [dependencies]
+# Ideally we should use the re-exported version of Arrow in DataFusion,
+# but here we depend on the Arrow crates directly so that this crate can be built without DataFusion.
+arrow-buffer = { workspace = true }
+arrow-schema = { workspace = true }
 thiserror = { workspace = true }
 serde = { workspace = true }
 serde_json = { workspace = true }
 chrono = { workspace = true }
 chrono-tz = { workspace = true }
-ryu = { workspace = true }
-datafusion = { workspace = true }
-datafusion-common = { workspace = true }
-datafusion-expr = { workspace = true }
 glob = { workspace = true }
-lexical-core = { workspace = true }
 num_enum = { workspace = true }
 figment = { workspace = true }
-futures = { workspace = true }
 half = { workspace = true }
 log = { workspace = true }
 iana-time-zone = { workspace = true }
