+++ crates/sail-common/src/string.rs
