+++ crates/sail-plan/Cargo.toml
@@ -5,6 +5,7 @@ edition = { workspace = true }
 
 [dependencies]
 sail-common = { path = "../sail-common" }
+sail-common-datafusion = { path = "../sail-common-datafusion" }
 sail-python-udf = { path = "../sail-python-udf" }
 
 datafusion = { workspace = true }
@@ -21,14 +22,11 @@ async-recursion = { workspace = true }
 lazy_static = { workspace = true }
 regex = { workspace = true }
 serde = { workspace = true }
-serde_json = { workspace = true }
 serde_arrow = { workspace = true }
-pyo3 = { workspace = true }
 futures = { workspace = true }
 comfy-table = { workspace = true }
 html-escape = { workspace = true }
 chrono = { workspace = true }
-chrono-tz = { workspace = true }
 ryu = { workspace = true }
 either = { workspace = true }
 num-traits = { workspace = true }
@@ -41,7 +39,6 @@ object_store = { workspace = true }
 aws-config = { workspace = true }
 aws-credential-types = { workspace = true }
 hdfs-native-object-store = { workspace = true, optional = true }
-log = { workspace = true }
 twox-hash = { workspace = true }
 aes = { workspace = true }
 aes-gcm = { workspace = true }
