+++ crates/sail-python-udf/Cargo.toml
@@ -5,13 +5,13 @@ edition = { workspace = true }
 
 [dependencies]
 sail-common = { path = "../sail-common" }
+sail-common-datafusion = { path = "../sail-common-datafusion" }
 
 thiserror = { workspace = true }
 datafusion = { workspace = true }
 datafusion-common = { workspace = true }
 datafusion-expr = { workspace = true }
 pyo3 = { workspace = true }
-async-trait = { workspace = true }
 futures = { workspace = true }
 tokio = { workspace = true }
 tokio-stream = { workspace = true }
