+++ crates/sail-server/Cargo.toml
@@ -11,7 +11,5 @@ tokio = { workspace = true }
 tonic = { workspace = true }
 tonic-reflection = { workspace = true }
 tonic-health = { workspace = true }
-tonic-types = { workspace = true }
 tower = { workspace = true }
-tower-http = { workspace = true }
 log = { workspace = true }
