+++ crates/sail-spark-connect/src/error.rs
@@ -73,7 +73,6 @@ impl SparkError {
 impl From<CommonError> for SparkError {
     fn from(error: CommonError) -> Self {
         match error {
-            CommonError::DataFusionError(e) => SparkError::DataFusionError(e),
             CommonError::MissingArgument(message) => SparkError::MissingArgument(message),
             CommonError::InvalidArgument(message) => SparkError::InvalidArgument(message),
             CommonError::NotSupported(message) => SparkError::NotSupported(message),
@@ -85,7 +84,6 @@ impl From<CommonError> for SparkError {
 impl From<SqlError> for SparkError {
     fn from(error: SqlError) -> Self {
         match error {
-            SqlError::DataFusionError(e) => SparkError::DataFusionError(e),
             SqlError::MissingArgument(message) => SparkError::MissingArgument(message),
             SqlError::InvalidArgument(message) => SparkError::InvalidArgument(message),
             SqlError::NotSupported(message) => SparkError::NotSupported(message),
